refactor useful form interactions for testing
